chore: remove mdx from docs build
We will be updating to mdx@2 soon which is esm only, but built in to
Gatsby. Due to the new documentation build process here and in
`npm/documentation`, there is much less need for us to verify all our
docs here. Therefore, it is best to remove this dependency now to avoid
needing to update it here at the same time as the docs site.